requirements.txt
llama_cpp_python==0.2.83
streamlit==1.37.0
googlesearch-python==1.2.4
nltk==3.8.1
rake_nltk==1.0.6
boilerpy3==1.0.7
